Add ex12_1 and ex13

+++ ex12.rb
@@ -1,4 +1,5 @@
 print "Give me a number: "
+# Turns input into a integer
 number = gets.chomp.to_i
 
 bigger = number * 100
@@ -6,7 +7,9 @@ puts "A bigger number is #{bigger}."
 
 print "Give me another number: "
 another = gets.chomp
-number = another.to_i
+
+# Turns number into a floating point
+number = another.to_f
 
 smaller = number / 100
 puts "A smaller number is #{smaller}."
